# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OF.ru",
        "version": "1.0.0"
    },
    "paths": {
...
        "/v1/articles/{slug}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "description": "Display a single item by slug.",
                "operationId": "c5a203fee5e8301ec6f0ca47073fb954",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "The slug of the item",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ArticleResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Resource not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ArticleCategoryResource": {
                "description": "ArticleCategory",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ArticleResource": {
                "description": "Article",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "likes": {
                        "type": "integer"
                    },
                    "dislikes": {
                        "type": "integer"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "content": {
                        "type": "integer"
                    },
                    "sources": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "link": {
                                    "type": "string"
                                },
                                "content": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "hyperlinks": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "category": {
                        "$ref": "#/components/schemas/ArticleCategoryResource"
                    },
                    "creator": {
                        "$ref": "#/components/schemas/AuthorResource"
                    },
                    "updater": {
                        "$ref": "#/components/schemas/AuthorResource"
                    }
                },
                "type": "object"
            },
            "AuthorResource": {
                "description": "Author",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "avatar": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/",
            "description": "Server parameters"
        }
    ]
}